+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.9] - Unreleased
+
+### Fixed
+
+- If a wrong password is entered for the SSH server connection, you will now be prompted to retry
+- Fixed several critical issues related to SSH command submissions
+
+----
+
 ## [0.3.8] - 2019-05-13
 
 ### Fixed